fold Domains into Scanners
