+++ src/os2datascanner/engine/scanners/scanner_types/exchange_scanner.py
@@ -1,21 +1,7 @@
+from os2datascanner.projects.admin.adminapp.models.scans.exchangescan_model import ExchangeScan
 from .scanner import Scanner
 
 
 class ExchangeScanner(Scanner):
-
-    def get_domain_urls(self):
-        """Return a list of valid domain urls."""
-        domains = []
-        for d in self.valid_domains:
-            domains.append(d.exchangedomain.dir_to_scan)
-        return domains
-
-    def get_domain_objects(self):
-        """
-        Returns a list of valid domain objects
-        :return: domain list
-        """
-        domains = []
-        for domain in self.valid_domains:
-            domains.append(domain.exchangedomain)
-        return domains
+    def get_domain_url(self):
+        return self.get_scanner_object().dir_to_scan
