+++ src/os2datascanner/projects/admin/adminapp/models/__init__.py
@@ -1,2 +1,25 @@
 # Import needed here for django models:
-
+from . import authentication_model  # noqa
+from . import authenticationmethods_model  # noqa
+from . import conversionqueueitem_model  # noqa
+from . import group_model  # noqa
+from . import match_model  # noqa
+from . import organization_model  # noqa
+from . import referrerurl_model  # noqa
+from . import regexpattern_model  # noqa
+from . import regexrule_model  # noqa
+from . import sensitivity_level  # noqa
+from . import statistic_model  # noqa
+from . import summary_model  # noqa
+from . import urllastmodified_model  # noqa
+from . import userprofile_model  # noqa
+from . import version_model  # noqa
+from . import webversion_model  # noqa
+from .scannerjobs import exchangescanner_model  # noqa
+from .scannerjobs import filescanner_model  # noqa
+from .scannerjobs import scanner_model  # noqa
+from .scannerjobs import webscanner_model  # noqa
+from .scans import exchangescan_model  # noqa
+from .scans import filescan_model  # noqa
+from .scans import scan_model  # noqa
+from .scans import webscan_model  # noqa
