+++ src/os2datascanner/projects/admin/adminapp/models/scannerjobs/scanner_model.py
@@ -19,6 +19,7 @@
 import os
 import datetime
 import json
+import re
 
 from django.core.validators import validate_comma_separated_integer_list
 from django.db import models
@@ -27,6 +28,7 @@ from django.contrib.postgres.fields import JSONField
 from model_utils.managers import InheritanceManager
 from recurrence.fields import RecurrenceField
 
+from ..authentication_model import Authentication
 from ..organization_model import Organization
 from ..group_model import Group
 from ..regexrule_model import RegexRule
@@ -111,6 +113,43 @@ class Scanner(models.Model):
     address_replace_text = models.CharField(max_length=2048, null=True,
                                             blank=True)
 
+    VALID = 1
+    INVALID = 0
+
+    validation_choices = (
+        (INVALID, "Ugyldig"),
+        (VALID, "Gyldig"),
+    )
+
+    url = models.CharField(max_length=2048, blank=False, verbose_name='URL')
+
+    authentication = models.OneToOneField(Authentication,
+                                          null=True,
+                                          related_name='%(app_label)s_%(class)s_authentication',
+                                          verbose_name='Brugernavn',
+                                          on_delete=models.SET_NULL)
+
+    validation_status = models.IntegerField(choices=validation_choices,
+                                            default=INVALID,
+                                            verbose_name='Valideringsstatus')
+
+    exclusion_rules = models.TextField(blank=True,
+                                       default="",
+                                       verbose_name='Ekskluderingsregler')
+
+    def exclusion_rule_list(self):
+        """Return the exclusion rules as a list of strings or regexes."""
+        REGEX_PREFIX = "regex:"
+        rules = []
+        for line in self.exclusion_rules.splitlines():
+            line = line.strip()
+            if line.startswith(REGEX_PREFIX):
+                rules.append(re.compile(line[len(REGEX_PREFIX):],
+                                        re.IGNORECASE))
+            else:
+                rules.append(line)
+        return rules
+
     @property
     def is_running(self) -> bool:
         '''Are any scans currently running against this scanner?'''
@@ -130,10 +169,6 @@ class Scanner(models.Model):
     ALREADY_RUNNING = (
         "Scanneren kunne ikke startes," +
         " fordi der allerede er en scanning i gang for den."
-    )
-    NO_VALID_DOMAINS = (
-        "Scanneren kunne ikke startes," +
-        " fordi den ikke har nogen gyldige domæner."
     )
     EXCHANGE_EXPORT_IS_RUNNING = (
         "Scanneren kunne ikke startes," +
@@ -160,10 +195,6 @@ class Scanner(models.Model):
             hour=Scanner.FIRST_START_TIME.hour + added_hours,
             minute=Scanner.FIRST_START_TIME.minute + added_minutes
         )
-
-    @property
-    def has_valid_domains(self):
-        return self.organization.os2webscanner_domain_organization.filter(validation_status=True).exists()
 
     @classmethod
     def modulo_for_starttime(cls, time):
@@ -197,9 +228,6 @@ class Scanner(models.Model):
         """
         if self.is_running:
             return Scanner.ALREADY_RUNNING
-
-        if not self.has_valid_domains:
-            return Scanner.NO_VALID_DOMAINS
 
         # Create a new Scan
         scan = self.create_scan()
