+++ src/os2datascanner/projects/admin/adminapp/templates/index.html
@@ -24,22 +24,6 @@ OS2datascanner
       </a>
     </li>
     <li class="col-xs-12 col-sm-6 col-md-4">
-      {% if settings.ENABLE_WEBSCAN %}
-        <a href="/webdomains/" class="panel panel-default">
-      {% elif settings.ENABLE_FILESCAN %}
-          <a href="/filedomains/" class="panel panel-default">
-      {% elif settings.ENABLE_EXCHANGESCAN %}
-          <a href="/exchangedomains/" class="panel panel-default">
-      {% endif %}
-        <div class="panel-body">
-          <span class="h2">
-            Domæner
-          </span>
-          <span class="glyphicon glyphicon-globe" aria-hidden="true"></span>
-        </div>
-      </a>
-    </li>
-    <li class="col-xs-12 col-sm-6 col-md-4">
       <a href="/rules/" class="panel panel-default">
         <div class="panel-body">
           <span class="h2">
