--- src/os2datascanner/projects/admin/adminapp/templates/os2datascanner/domain_above_content.html
@@ -1,12 +1,0 @@
-<p class="lead">Indeholder de filer eller ressourcer der skal gennemgås.</p>
-<ul class="nav nav-tabs" role="tablist">
-    {% if settings.ENABLE_WEBSCAN %}
-    <li role="presentation"{% if active == 'web' %} class="active"{% endif %}><a href="/webdomains/" role="tab">Webdomæner</a></li>
-    {% endif %}
-    {% if settings.ENABLE_FILESCAN %}
-        <li role="presentation"{% if active == 'file' %} class="active"{% endif %}><a href="/filedomains/" role="tab">Fildomæner</a></li>
-    {% endif %}
-    {% if settings.ENABLE_EXCHANGESCAN %}
-    <li role="presentation"{% if active == 'exchange' %} class="active"{% endif %}><a href="/exchangedomains/" role="tab">Exchangedomæner</a></li>
-    {% endif %}
-</ul>
