--- src/os2datascanner/projects/admin/adminapp/templates/os2datascanner/domain_confirm_delete.html
@@ -1,20 +1,0 @@
-{% extends "os2datascanner/base.html" %}
-
-{% block rendertype %}regular{% endblock %}
-{% block heading %}
-Slet {{ object }}?
-{% endblock %}
-
-{% block body %}
-  <div class="container">
-    <p>
-      Ønsker du virkelig at slette dette domæne? Handlingen kan ikke
-      fortrydes.
-    </p>
-    <form action="" method="post">{% csrf_token %}
-      {{ form.as_p }}
-      <input class="btn btn-primary" type="submit" value="Ja" />
-      <a class="btn btn-default" href="javascript:history.go(-1)">Nej</a>
-    </form>
-  </div>
-{% endblock %}
