+++ src/os2datascanner/projects/admin/adminapp/templates/os2datascanner/domain_form.html
@@ -163,23 +163,6 @@
       </div>
   {% endif %}
 
-  {% if form.userlist %}
-      <div id="{{ form.userlist.auto_id }}_container" class="col-sm-12{% if form.userlist.errors %} has-error{% endif %}">
-          <div class="form-group">
-              <label class="control-label col-sm-3" for="id_{{ form.userlist.name }}">Liste af mail brugere</label>
-              <div class="col-sm-9">
-                  <input type="file" class="form-control" name="{{ form.userlist.name }}" id="id_{{ form.userlist.name }}" {% if form.userlist.value %}value="{{ form.userlist.value }}"{% endif %}>
-                  {% if form.userlist.help_text %}
-                      <p>
-                          <small>{{ form.userlist.help_text }}</small>
-                      </p>
-                  {% endif %}
-                  {% if form.userlist.errors %}{{ form.userlist.errors }}{% endif %}
-              </div>
-          </div>
-      </div>
-  {% endif %}
-
   {% if form.group %}
     <div id="{{ form.group.auto_id }}_container" class="col-sm-12{% if form.group.errors %} has-error{% endif %}">
       <div class="form-group">
