+++ src/os2datascanner/projects/admin/adminapp/templates/os2datascanner/scanner_above_content.html
@@ -1,4 +1,7 @@
-<p class="lead">Gennemgår indholdet af domænerne med de opsatte regler.</p>
+<p class="lead">
+    Angiver hvilket domæne der skal gennemgås samt hvilke regler der skal
+    tjekkes.
+</p>
 <ul class="nav nav-tabs" role="tablist">
     {% if settings.ENABLE_WEBSCAN %}
     <li role="presentation"{% if active == 'web' %} class="active"{% endif %}><a href="/webscanners/" role="tab">Webscannere</a></li>
