+++ src/os2datascanner/projects/admin/adminapp/urls.py
@@ -24,11 +24,8 @@ from .models.scans.scan_model import Scan
 from .models.scannerjobs.exchangescanner_model import ExchangeScanner
 from .models.scannerjobs.filescanner_model import FileScanner
 from .models.scannerjobs.webscanner_model import WebScanner
-from .views.exchangedomain_views import ExchangeDomainList, ExchangeDomainCreate, ExchangeDomainUpdate, \
-    ExchangeDomainDelete
 from .views.exchangescanner_views import ExchangeScannerList, ExchangeScannerCreate, ExchangeScannerUpdate, \
     ExchangeScannerDelete, ExchangeScannerRun, ExchangeScannerAskRun
-from .views.filedomain_views import FileDomainList, FileDomainCreate, FileDomainUpdate, FileDomainDelete
 from .views.filescanner_views import FileScannerCreate, FileScannerRun, FileScannerAskRun, FileScannerUpdate, \
     FileScannerDelete, FileScannerList
 from .views.report_views import ScanReportLog, CSVReportDetails, ReportDetails, ReportList, ReportDelete
@@ -39,10 +36,10 @@ from .views.views import OrganizationUpdate, OrganizationList
 from .views.views import SummaryList, SummaryCreate, SummaryUpdate, SummaryDelete
 from .views.views import SummaryReport, DialogSuccess, SystemStatusView
 from .views.views import file_upload, referrer_content
-from .views.webdomain_views import WebDomainList, WebDomainCreate, WebDomainUpdate, WebDomainDelete, WebDomainValidate
-from .views.webscanner_views import WebScannerCreate, WebScannerUpdate, WebScannerDelete, WebScannerRun, \
-    WebScannerAskRun, WebScannerList
-
+from .views.webscanner_views import (WebScannerCreate, WebScannerUpdate,
+                                     WebScannerDelete, WebScannerRun,
+                                     WebScannerAskRun, WebScannerList,
+                                     WebScannerValidate)
 
 urlpatterns = [
     # App URLs
@@ -64,6 +61,8 @@ urlpatterns = [
     url(r'^webscanners/add/$', WebScannerCreate.as_view(), name='webscanner_add'),
     url(r'^webscanners/(?P<pk>\d+)/delete/$', WebScannerDelete.as_view(),
         name='scanner_delete'),
+    url(r'^webscanners/(?P<pk>\d+)/validate/$', WebScannerValidate.as_view(),
+        name='web_scanner_validate'),
     url(r'^webscanners/(?P<pk>\d+)/run/$', WebScannerRun.as_view(),
         name='webscanner_run'),
     url(r'^webscanners/(?P<pk>\d+)/askrun/$',
@@ -86,29 +85,6 @@ urlpatterns = [
             template_name='os2datascanner/scanner_askrun.html',
             model=FileScanner),
         name='filescanner_askrun'),
-    url(r'^exchangedomains/$', ExchangeDomainList.as_view(), name='exchangedomains'),
-    url(r'^exchangedomains/add/$', ExchangeDomainCreate.as_view(), name='exchangedomain_add'),
-    url(r'^(exchangedomains)/(\d+)/(success)/$', DialogSuccess.as_view()),
-    url(r'^exchangedomains/(?P<pk>\d+)/$', ExchangeDomainUpdate.as_view(),
-        name='exchange_domain_update'),
-    url(r'^exchangedomains/(?P<pk>\d+)/delete/$', ExchangeDomainDelete.as_view(),
-        name='exchange_domain_delete'),
-    url(r'^filedomains/$', FileDomainList.as_view(), name='filedomains'),
-    url(r'^filedomains/add/$', FileDomainCreate.as_view(), name='filedomain_add'),
-    url(r'^(filedomains)/(\d+)/(success)/$', DialogSuccess.as_view()),
-    url(r'^filedomains/(?P<pk>\d+)/$', FileDomainUpdate.as_view(),
-        name='file_domain_update'),
-    url(r'^filedomains/(?P<pk>\d+)/delete/$', FileDomainDelete.as_view(),
-        name='file_domain_delete'),
-    url(r'^webdomains/$', WebDomainList.as_view(), name='webdomains'),
-    url(r'^webdomains/add/$', WebDomainCreate.as_view(), name='webdomain_add'),
-    url(r'^webdomains/(?P<pk>\d+)/validate/$', WebDomainValidate.as_view(),
-        name='web_domain_validate'),
-    url(r'^(webdomains)/(\d+)/(success)/$', DialogSuccess.as_view()),
-    url(r'^webdomains/(?P<pk>\d+)/$', WebDomainUpdate.as_view(),
-        name='web_domain_update'),
-    url(r'^webdomains/(?P<pk>\d+)/delete/$', WebDomainDelete.as_view(),
-        name='web_domain_delete'),
     url(r'^rules/$', RuleList.as_view(), name='rules'),
     url(r'^rules/add/$', RuleCreate.as_view(), name='rule_add'),
     url(r'^rules/(?P<pk>\d+)/$', RuleUpdate.as_view(),
@@ -189,19 +165,19 @@ urlpatterns = [
         ),
 
     # General dialog success handler
-    url(r'^(webscanners|webdomains|rules|groups|reports/summaries)/(\d+)/(created)/$',
+    url(r'^(webscanners|rules|groups|reports/summaries)/(\d+)/(created)/$',
         DialogSuccess.as_view()),
-    url(r'^(webscanners|webdomains|rules|groups|reports/summaries)/(\d+)/(saved)/$',
-        DialogSuccess.as_view()),
-    # General dialog success handler
-    url(r'^(filescanners|filedomains|rules|groups|reports/summaries)/(\d+)/(created)/$',
-        DialogSuccess.as_view()),
-    url(r'^(filescanners|filedomains|rules|groups|reports/summaries)/(\d+)/(saved)/$',
+    url(r'^(webscanners|rules|groups|reports/summaries)/(\d+)/(saved)/$',
         DialogSuccess.as_view()),
     # General dialog success handler
-    url(r'^(exchangescanners|exchangedomains|rules|groups|reports/summaries)/(\d+)/(created)/$',
+    url(r'^(filescanners|rules|groups|reports/summaries)/(\d+)/(created)/$',
         DialogSuccess.as_view()),
-    url(r'^(exchangescanners|exchangedomains|rules|groups|reports/summaries)/(\d+)/(saved)/$',
+    url(r'^(filescanners|rules|groups|reports/summaries)/(\d+)/(saved)/$',
+        DialogSuccess.as_view()),
+    # General dialog success handler
+    url(r'^(exchangescanners|rules|groups|reports/summaries)/(\d+)/(created)/$',
+        DialogSuccess.as_view()),
+    url(r'^(exchangescanners|rules|groups|reports/summaries)/(\d+)/(saved)/$',
         DialogSuccess.as_view()),
     url(r'^jsi18n/$', JavaScriptCatalog.as_view(
         packages=('os2webscanner', 'recurrence'),
