+++ src/os2datascanner/projects/admin/adminapp/views/domain_views.py
@@ -3,7 +3,6 @@ from django import forms
 from .views import RestrictedListView, RestrictedCreateView, \
     RestrictedUpdateView, RestrictedDetailView
 from ..models.authentication_model import Authentication
-from ..models.domains.domain_model import Domain
 from ..utils import domain_form_manipulate
