+++ src/os2datascanner/projects/admin/adminapp/views/exchangescanner_views.py
@@ -1,3 +1,5 @@
+from django import forms
+
 from .scanner_views import *
 from ..models.scannerjobs.exchangescanner_model import ExchangeScanner
 
@@ -13,7 +15,7 @@ class ExchangeScannerCreate(ScannerCreate):
     """Create a exchange scanner view."""
 
     model = ExchangeScanner
-    fields = ['name', 'schedule', 'domains',
+    fields = ['name', 'schedule',
               'do_name_scan', 'do_ocr', 'do_address_scan', 'do_last_modified_check',
               'regex_rules', 'recipients']
 
@@ -21,18 +23,52 @@ class ExchangeScannerCreate(ScannerCreate):
         """The URL to redirect to after successful creation."""
         return '/exchangescanners/%s/created/' % self.object.pk
 
+    def get_form(self, form_class=None):
+        """Adds special field password."""
+        if form_class is None:
+            form_class = self.get_form_class()
+
+        form = super().get_form(form_class)
+        form.fields['username'] = forms.CharField(max_length=1024, required=False)
+        form.fields['password'] = forms.CharField(max_length=50, required=False)
+        return form
 
 class ExchangeScannerUpdate(ScannerUpdate):
     """Update a scanner view."""
 
     model = ExchangeScanner
-    fields = ['name', 'schedule', 'domains',
+    fields = ['name', 'schedule',
               'do_name_scan', 'do_ocr', 'do_address_scan', 'do_last_modified_check',
-              'regex_rules', 'recipients']
+              'regex_rules', 'recipients', 'url', 'userlist']
 
     def get_success_url(self):
-        """The URL to redirect to after successful update."""
-        return '/exchangescanners/%s/saved/' % self.object.pk
+        """The URL to redirect to after successful updating.
+
+        Will redirect the user to the validate view if the form was submitted
+        with the 'save_and_validate' button.
+        """
+        if 'save_and_validate' in self.request.POST:
+            return 'validate/'
+        else:
+            return '/exchangescanners/%s/saved/' % self.object.pk
+
+    def get_form(self, form_class=None):
+        """Adds special field password and decrypts password."""
+        if form_class is None:
+            form_class = self.get_form_class()
+
+        form = super().get_form(form_class)
+        exchangedomain = self.get_object()
+        authentication = exchangedomain.authentication
+        form.fields['username'] = forms.CharField(max_length=1024, required=False)
+        form.fields['password'] = forms.CharField(max_length=50, required=False)
+        if authentication.username:
+            form.fields['username'].initial = authentication.username
+        if authentication.ciphertext:
+            password = decrypt(bytes(authentication.iv),
+                               bytes(authentication.ciphertext))
+            form.fields['password'].initial = password
+        return form
 
 
 class ExchangeScannerDelete(ScannerDelete):
