+++ src/os2datascanner/projects/admin/adminapp/views/views.py
@@ -34,9 +34,7 @@ from shutil import copyfile
 
 from ..forms import FileUploadForm
 from ..models.conversionqueueitem_model import ConversionQueueItem
-from ..models.domains.exchangedomain_model import ExchangeDomain
 from ..models.scannerjobs.exchangescanner_model import ExchangeScanner
-from ..models.domains.filedomain_model import FileDomain
 from ..models.scannerjobs.filescanner_model import FileScanner
 from ..models.group_model import Group
 from ..models.organization_model import Organization
@@ -45,7 +43,6 @@ from ..models.regexrule_model import RegexRule
 from ..models.scans.scan_model import Scan
 from ..models.summary_model import Summary
 from ..models.userprofile_model import UserProfile
-from ..models.domains.webdomain_model import WebDomain
 from ..models.scannerjobs.webscanner_model import WebScanner
 from ..utils import scans_for_summary_report, do_scan, as_file_uri
 
@@ -138,11 +135,13 @@ class OrganizationList(RestrictedListView):
             def top_level(d):
                 return '.'.join(d.strip('/').split('.')[-2:])
 
-            tlds = set([top_level(d.url) for d in org.os2webscanner_domain_organization.all()])
+            tlds = set([top_level(d.url) for d in
+                        org.scanners.all()])
 
             for tld in tlds:
                 sub_domains = [
-                    d.url for d in org.os2webscanner_domain_organization.all() if top_level(d.url) == tld
+                    d.url for d in org.scanners.all() if top_level(d.url) ==
+                                                         tld
                 ]
                 tld_list.append({'tld': tld, 'domains': sub_domains})
 
@@ -417,11 +416,8 @@ class DialogSuccess(TemplateView):
     template_name = 'os2datascanner/dialogsuccess.html'
 
     type_map = {
-        'webdomains': WebDomain,
         'webscanners': WebScanner,
-        'filedomains': FileDomain,
         'filescanners': FileScanner,
-        'exchangedomains': ExchangeDomain,
         'exchangescanners': ExchangeScanner,
         'rules': RegexRule,
         'groups': Group,
@@ -665,7 +661,7 @@ def file_upload(request):
             #
             if not isinstance(scan, Scan):
                 raise RuntimeError("Unable to perform scan - check user has"
-                                   "organization and valid domain")
+                                   "organization and valid scanner")
             # We now have the scan object
             if not params['output_spreadsheet_file']:
                 return HttpResponseRedirect('/report/{0}/'.format(scan.pk))
